refactoring type
